Created a exam insertion view, also now TodayView can display exams i…

+++ app/src/main/java/net/felipealafy/studentplanner/datamodels/Exam.kt
@@ -9,7 +9,7 @@ data class Exam(
     val subjectId: String,
     val name: String,
     val grade: Float,
-    val gradeWeight: Int,
+    val gradeWeight: Float,
     val start: LocalDateTime,
     val end: LocalDateTime,
 )
